Creat htmlgui file, addnode api, support p2sh address display
