+++ src/net.c
@@ -347,20 +347,18 @@ void btc_node_group_event_loop(btc_node_group* group)
     event_base_dispatch(group->event_base);
 }
 
-void btc_node_group_add_node(btc_node_group* group, btc_node* node)
+btc_node *btc_node_group_add_node(btc_node_group* group, btc_node* node)
 {
     size_t j;
     for ( j = 0; j < group->nodes->len; j++) {
         btc_node* existing_node = vector_idx(group->nodes, j);
         if ( memcmp(&((struct sockaddr_in*)&existing_node->addr)->sin_addr, &((struct sockaddr_in*)&node->addr)->sin_addr, sizeof(&((struct sockaddr_in*)&existing_node->addr)->sin_addr)) == 0 ) 
-            break;
+            return(existing_node);
     }
-    if ( j == group->nodes->len )
-    {
-        vector_add(group->nodes, node);
-        node->nodegroup = group;
-        node->nodeid = group->nodes->len;
-    }
+    vector_add(group->nodes, node);
+    node->nodegroup = group;
+    node->nodeid = group->nodes->len;
+    return(node);
 }
 
 int btc_node_group_amount_of_connected_nodes(btc_node_group* group, enum NODE_STATE state)
@@ -613,10 +611,12 @@ btc_bool btc_node_group_add_peers_by_ip_or_seed(btc_node_group *group, const cha
 
             /* create a node */
             btc_node* node = btc_node_new();
-            if (btc_node_set_ipport(node, ip) > 0) {
+            if (btc_node_set_ipport(node, ip) > 0)
+            {
                 /* add the node to the group */
-                btc_node_group_add_node(group, node);
-            }
+                if ( btc_node_group_add_node(group, node) != node )
+                    btc_node_free(node);
+            } else btc_node_free(node);
         }
         vector_free(ips_dns, true);
     } else {
@@ -630,9 +630,11 @@ btc_bool btc_node_group_add_peers_by_ip_or_seed(btc_node_group *group, const cha
                 
                 sprintf(ipaddr,"%s:%u",working_str,group->chainparams->default_port);
                 //fprintf(stderr,"setnode.(%s) -> %s\n",working_str,ipaddr);
-                if (btc_node_set_ipport(node, ipaddr) > 0) {
-                    btc_node_group_add_node(group, node);
-                }
+                if ( btc_node_set_ipport(node, ipaddr) > 0)
+                {
+                    if (btc_node_group_add_node(group, node) != node )
+                        btc_node_free(node);
+                } else btc_node_free(node);
                 offset = 0;
                 memset(working_str, 0, sizeof(working_str));
             } else if (ips[i] != ' ' && offset < sizeof(working_str)) {
