+++ src/netspv.c
@@ -396,15 +396,17 @@ void btc_net_spv_post_cmd(btc_node *node, btc_p2p_msg_hdr *hdr, struct const_buf
                 deser_bytes(ipdata,buf,sizeof(ipdata));
                 deser_u16(&port,buf);
                 expand_ipbits(ipaddr,*(uint32_t *)&ipdata[12]);
-                if ( 0 && (services & NODE_NSPV) != 0 )
+                if ( 1 && (services & NODE_NSPV) != 0 )
                 {
                     revport = ((port >> 8) & 0xff) | ((port & 0xff) << 8);
                     sprintf(ipaddr+strlen(ipaddr),":%u",revport);
                     //fprintf(stderr,"%d: %u %llx %s\n",i,timestamp,(long long)services,ipaddr);
                     tmpnode = btc_node_new();
                     if ( btc_node_set_ipport(tmpnode,ipaddr) > 0 )
-                        btc_node_group_add_node(node->nodegroup,tmpnode);
-                    else btc_node_free(tmpnode);
+                    {
+                        if (btc_node_group_add_node(node->nodegroup,tmpnode) != tmpnode )
+                            btc_node_free(tmpnode);
+                    } else btc_node_free(tmpnode);
                 }
             }
             node->gotaddrs = (uint32_t)time(NULL);
